Implement cascade param and HTTP `OPTIONS` handling
Introduce config to set the IPNI cascade label with default of
`ipfs-dht`, and optionally require `?cascade=<label>` on IPNI HTTP
lookup APIs before responding with results. The rationale for making the
query parameter check optional is to offer an easier out-of-the-box
setup and demo.

Add handling for HTTP `OPTIONS` requests according to the IPNI
specification. While at it, also handle such requests on HTTP delegated
routing API.